+++ CHANGES.md
@@ -14,6 +14,7 @@ Core Changes:
 
 Language Improvements:
 
+- enh(csharp) Add generic modifiers (in, out) (#2378) [Josh Goebel][]
 - (fortran) Add Fortran 2018 keywords and coarray intrinsics (#2361) [Sam Miller][]
 - (delphi) highlight hexadecimal, octal, and binary numbers (#2370) [Robert Riebisch]()
 - enh(plaintext) added `text` and `txt` as alias (#2360) [Taufik Nurrohman][]
@@ -22,6 +23,7 @@ Developer Tools:
 
 - none.
 
+[Josh Goebel]: https://github.com/yyyc514
 [Sam Miller]: https://github.com/smillerc
 [Robert Riebisch]: https://github.com/bttrx
 [Taufik Nurrohman]: https://github.com/taufik-nurrohman